+++ packages/reactabular-virtualized/src/body-row.js
@@ -6,12 +6,19 @@ class BodyRow extends React.Component {
     super(props);
 
     this.ref = null;
+
+    this.updateHeight = this.updateHeight.bind(this);
   }
   componentDidMount() {
-    this.context.updateHeight(
-      this.props['data-rowkey'],
-      this.ref.offsetHeight
-    );
+    this.updateHeight();
+  }
+  componentDidUpdate() {
+    // Capture height data only during initial measurement for performance.
+    // This loses some accuracy if row height changes, but it's good enough
+    // for most purposes.
+    if (this.context.initialMeasurement) {
+      this.updateHeight();
+    }
   }
   render() {
     return React.createElement(
@@ -26,8 +33,17 @@ class BodyRow extends React.Component {
       }
     );
   }
+  updateHeight() {
+    this.context.updateHeight(
+      this.props['data-rowkey'],
+      this.ref.offsetHeight
+    );
+  }
 }
 BodyRow.contextTypes = bodyRowContextTypes;
 BodyRow.propTypes = bodyRowTypes;
+
+// Tell reactabular-table to update anyway.
+BodyRow.shouldComponentUpdate = true;
 
 export default BodyRow;